V22F-493 Tick to ulong

+++ SDALib-CPP/SDALib-CPP/SDAData.hpp
@@ -4,12 +4,12 @@
 struct SDAData
 {
 public:
-    SDAData(tCarElt* p_car, tSituation* p_situation, int p_tickCount) : Car(*p_car), Situation(*p_situation), TickCount(p_tickCount) {}
+    SDAData(tCarElt* p_car, tSituation* p_situation, unsigned long p_tickCount) : Car(*p_car), Situation(*p_situation), TickCount(p_tickCount) {}
     SDAData()
     {
         TickCount = 0;
     }
     tCarElt Car;
     tSituation Situation;
-    int TickCount;
+    unsigned long TickCount;
 };
